reorganized agent code
* reorganized agent code into multiple distinct files
* cleaned up some prompt details
